fixed code smells
